requirements.txt
streamlit==1.45.1
pandas==2.2.3
openpyxl==3.1.5
streamlit-authenticator==0.4.2
PyYAML==6.0.2
pydantic==2.11.5
scikit-learn==1.6.1
numpy==2.2.3
joblib==1.5.0
altair==5.5.0